``` refactor(api): reorganize persistence sdk package and class names
Refactor the persistence SDK by introducing dedicated 'call' and 'response'
sub-packages. Rename API call classes with a 'Call' suffix and response
objects with an 'Object' suffix for improved clarity and consistency.
The base ApiResponse class is also moved to the new response package.
Added package-info.java for persistencesdk documentation.

This refactoring enhances the modularity and maintainability of the SDK,
making it easier to understand the purpose of each class within the
persistence layer and promoting clearer separation of concerns.
```